demo/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sse-starlette==2.0.0
pydantic==2.5.3
httpx==0.26.0
python-dotenv==1.0.0
openai==1.12.0
gradio_client>=0.10.0